+++ testing/simulated/simulated_test.go
@@ -30,7 +30,6 @@ import (
 
 	"github.com/berachain/beacon-kit/beacon/blockchain"
 	"github.com/berachain/beacon-kit/consensus/cometbft/service/encoding"
-	"github.com/berachain/beacon-kit/engine-primitives/errors"
 	"github.com/berachain/beacon-kit/log/phuslu"
 	"github.com/berachain/beacon-kit/node-core/components/signer"
 	"github.com/berachain/beacon-kit/primitives/common"
@@ -41,7 +40,6 @@ import (
 	"github.com/cometbft/cometbft/abci/types"
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
 	"github.com/spf13/viper"
-	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -180,215 +178,6 @@ func (s *SimulatedSuite) TestFullLifecycle_ValidBlock_IsSuccessful() {
 	)
 	s.Require().NoError(err)
 	s.Require().Equal(proposedBlock.Message.GetHeader().GetBodyRoot(), stateHeader.GetBodyRoot())
-}
-
-// TestProcessProposal_BadBlock_IsRejected effectively serves as a test for how a valid node would react to
-// a malicious proposer proposing a block with an invalid EVM transaction.
-func (s *SimulatedSuite) TestProcessProposal_BadBlock_IsRejected() {
-	const blockHeight = 1
-	const coreLoopIterations = 1
-
-	// Initialize the chain state.
-	s.initializeChain()
-
-	// Retrieve the BLS signer and proposer address.
-	blsSigner := simulated.GetBlsSigner(s.HomeDir)
-	pubkey, err := blsSigner.GetPubKey()
-	s.Require().NoError(err)
-
-	// Go through 1 iteration of the core loop to bypass any startup specific edge cases such as sync head on startup.
-	proposals := s.CoreLoop(blockHeight, coreLoopIterations, blsSigner)
-	s.Require().Len(proposals, coreLoopIterations)
-
-	// Prepare a valid block proposal.
-	proposalTime := time.Now()
-	proposal, err := s.SimComet.Comet.PrepareProposal(s.Ctx, &types.PrepareProposalRequest{
-		Height:          blockHeight + coreLoopIterations,
-		Time:            proposalTime,
-		ProposerAddress: pubkey.Address(),
-	})
-	s.Require().NoError(err)
-	s.Require().NotEmpty(proposal)
-
-	// Unmarshal the proposal block.
-	proposedBlock, err := encoding.UnmarshalBeaconBlockFromABCIRequest(
-		proposal.Txs,
-		blockchain.BeaconBlockTxIndex,
-		s.TestNode.ChainSpec.ActiveForkVersionForSlot(blockHeight+coreLoopIterations),
-	)
-	s.Require().NoError(err)
-
-	// Create a malicious block by injecting an invalid transaction.
-	maliciousBlock := simulated.CreateInvalidBlock(require.New(s.T()), proposedBlock, blsSigner, s.TestNode.ChainSpec, s.GenesisValidatorsRoot)
-	maliciousBlockBytes, err := maliciousBlock.MarshalSSZ()
-	s.Require().NoError(err)
-
-	// Replace the valid block with the malicious block in the proposal.
-	proposal.Txs[0] = maliciousBlockBytes
-
-	// Reset the log buffer to discard old logs we don't care about
-	s.LogBuffer.Reset()
-	// Process the proposal containing the malicious block.
-	processResp, err := s.SimComet.Comet.ProcessProposal(s.Ctx, &types.ProcessProposalRequest{
-		Txs:             proposal.Txs,
-		Height:          blockHeight + coreLoopIterations,
-		ProposerAddress: pubkey.Address(),
-		Time:            proposalTime,
-	})
-	s.Require().NoError(err)
-	s.Require().Equal(types.PROCESS_PROPOSAL_STATUS_REJECT, processResp.Status)
-
-	// Verify that the log contains the expected error message.
-	s.Require().Contains(s.LogBuffer.String(), errors.ErrInvalidPayloadStatus.Error())
-	// Note this error message may change across execution clients. Base fee changes with number of core loop iterations.
-	s.Require().Contains(s.LogBuffer.String(), "max fee per gas less than block base fee: address 0x71562b71999873DB5b286dF957af199Ec94617F7, maxFeePerGas: 10000000, baseFee: 765625000")
-}
-
-// TestProcessProposal_InvalidTimestamps_Errors effectively serves as a test for how a valid node would react to
-// a malicious proposer attempting to use a future timestamp in the block that does not match the consensus timestamp.
-func (s *SimulatedSuite) TestProcessProposal_InvalidTimestamps_Errors() {
-	const blockHeight = 1
-	const coreLoopIterations = 1
-
-	// Initialize the chain state.
-	s.initializeChain()
-
-	// Retrieve the BLS signer and proposer address.
-	blsSigner := simulated.GetBlsSigner(s.HomeDir)
-	pubkey, err := blsSigner.GetPubKey()
-	s.Require().NoError(err)
-
-	// Go through 1 iteration of the core loop to bypass any startup specific edge cases such as sync head on startup.
-	proposals := s.CoreLoop(blockHeight, coreLoopIterations, blsSigner)
-	s.Require().Len(proposals, coreLoopIterations)
-
-	// Prepare a valid block proposal, but 2 seconds in the future (i.e. attempt to roll timestamp forward)
-	correctTime := time.Now()
-	proposalTime := correctTime.Add(2 * time.Second)
-	maliciousProposal, err := s.SimComet.Comet.PrepareProposal(s.Ctx, &types.PrepareProposalRequest{
-		Height:          blockHeight + coreLoopIterations,
-		Time:            proposalTime,
-		ProposerAddress: pubkey.Address(),
-	})
-	s.Require().NoError(err)
-	s.Require().NotEmpty(maliciousProposal)
-
-	// Reset the log buffer to discard old logs we don't care about
-	s.LogBuffer.Reset()
-	// Process the proposal containing the malicious block.
-	processResp, err := s.SimComet.Comet.ProcessProposal(s.Ctx, &types.ProcessProposalRequest{
-		Txs:             maliciousProposal.Txs,
-		Height:          blockHeight + coreLoopIterations,
-		ProposerAddress: pubkey.Address(),
-		// Use the correct time as the actual consensus time, which mismatches the proposal time.
-		Time: correctTime,
-	})
-	s.Require().NoError(err)
-	s.Require().Equal(types.PROCESS_PROPOSAL_STATUS_REJECT, processResp.Status)
-	s.Require().Contains(s.LogBuffer.String(), "timestamp too far in the future")
-}
-
-// TestProcessProposal_CrashedExecutionClient_Errors effectively serves as a test for how a valid node would react to
-// a valid block being proposed but the execution client has crashed.
-func (s *SimulatedSuite) TestProcessProposal_CrashedExecutionClient_Errors() {
-	const blockHeight = 1
-	const coreLoopIterations = 1
-
-	// Initialize the chain state.
-	s.initializeChain()
-
-	// Retrieve the BLS signer and proposer address.
-	blsSigner := simulated.GetBlsSigner(s.HomeDir)
-	pubkey, err := blsSigner.GetPubKey()
-	s.Require().NoError(err)
-
-	// Go through 1 iteration of the core loop to bypass any startup specific edge cases such as sync head on startup.
-	proposals := s.CoreLoop(blockHeight, coreLoopIterations, blsSigner)
-	s.Require().Len(proposals, coreLoopIterations)
-
-	// Prepare a valid block proposal.
-	proposalTime := time.Now()
-	proposal, err := s.SimComet.Comet.PrepareProposal(s.Ctx, &types.PrepareProposalRequest{
-		Height:          blockHeight + coreLoopIterations,
-		Time:            proposalTime,
-		ProposerAddress: pubkey.Address(),
-	})
-	s.Require().NoError(err)
-	s.Require().NotEmpty(proposal)
-
-	// Reset the log buffer to discard old logs we don't care about.
-	s.LogBuffer.Reset()
-	// Kill the execution client.
-	err = s.ElHandle.Close()
-	s.Require().NoError(err)
-	// Process the proposal containing the malicious block.
-	processResp, err := s.SimComet.Comet.ProcessProposal(s.Ctx, &types.ProcessProposalRequest{
-		Txs:             proposal.Txs,
-		Height:          blockHeight + coreLoopIterations,
-		ProposerAddress: pubkey.Address(),
-		Time:            proposalTime,
-	})
-	s.Require().NoError(err)
-	s.Require().Equal(types.PROCESS_PROPOSAL_STATUS_REJECT, processResp.Status)
-	s.Require().Contains(s.LogBuffer.String(), "got an unexpected server error in JSON-RPC response failed to convert from jsonrpc.Error")
-}
-
-// TestFinalizeBlock_BadBlock_Errors effectively serves as a test for how a valid node would react to
-// a malicious consensus majority agreeing to a block with an invalid EVM transaction.
-func (s *SimulatedSuite) TestFinalizeBlock_BadBlock_Errors() {
-	const blockHeight = 1
-	const coreLoopIterations = 1
-
-	// Initialize the chain state.
-	s.initializeChain()
-
-	// Retrieve the BLS signer and proposer address.
-	blsSigner := simulated.GetBlsSigner(s.HomeDir)
-	pubkey, err := blsSigner.GetPubKey()
-	s.Require().NoError(err)
-
-	// Go through 1 iteration of the core loop to bypass any startup specific edge cases such as sync head on startup.
-	proposals := s.CoreLoop(blockHeight, coreLoopIterations, blsSigner)
-	s.Require().Len(proposals, coreLoopIterations)
-
-	// Prepare a valid block proposal.
-	proposalTime := time.Now()
-	proposal, err := s.SimComet.Comet.PrepareProposal(s.Ctx, &types.PrepareProposalRequest{
-		Height:          blockHeight + coreLoopIterations,
-		Time:            proposalTime,
-		ProposerAddress: pubkey.Address(),
-	})
-	s.Require().NoError(err)
-	s.Require().NotEmpty(proposal)
-
-	// Unmarshal the proposal block.
-	proposedBlock, err := encoding.UnmarshalBeaconBlockFromABCIRequest(
-		proposal.Txs,
-		blockchain.BeaconBlockTxIndex,
-		s.TestNode.ChainSpec.ActiveForkVersionForSlot(blockHeight+coreLoopIterations),
-	)
-	s.Require().NoError(err)
-
-	// Create a malicious block by injecting an invalid transaction.
-	maliciousBlock := simulated.CreateInvalidBlock(require.New(s.T()), proposedBlock, blsSigner, s.TestNode.ChainSpec, s.GenesisValidatorsRoot)
-	maliciousBlockBytes, err := maliciousBlock.MarshalSSZ()
-	s.Require().NoError(err)
-
-	// Replace the valid block with the malicious block in the proposal.
-	proposal.Txs[0] = maliciousBlockBytes
-
-	// Reset the log buffer to discard old logs we don't care about
-	s.LogBuffer.Reset()
-	// Finalize the proposal containing the malicious block.
-	finalizeResp, err := s.SimComet.Comet.FinalizeBlock(s.Ctx, &types.FinalizeBlockRequest{
-		Txs:             proposal.Txs,
-		Height:          blockHeight + coreLoopIterations,
-		ProposerAddress: pubkey.Address(),
-		Time:            proposalTime,
-	})
-	s.Require().ErrorIs(err, errors.ErrInvalidPayloadStatus)
-	s.Require().Nil(finalizeResp)
-	s.Require().Contains(s.LogBuffer.String(), "max fee per gas less than block base fee: address 0x71562b71999873DB5b286dF957af199Ec94617F7, maxFeePerGas: 10000000, baseFee: 765625000")
 }
 
 // CoreLoop will iterate through the core loop `iterations` times, i.e. Propose, Process, Finalize and Commit.
